requirements.txt
Flask==1.1.2
pandas
wikipedia
google-cloud-language